JIT: fix FitsIn<int32_t> assert in BitOperations.Rotate{Left,Right} const-fold
NI_PRIMITIVE_RotateLeft/Right's constant-folding path passes the unsigned
fold result directly to gtNewIconNode(ssize_t, TYP_INT). For TYP_INT/TYP_UINT
operands with the high bit set (e.g. RotateRight(0xFFFFFFFFu, k) which folds
to 0xFFFFFFFF), the implicit uint32_t-to-ssize_t conversion zero-extends to
a positive value (4294967295) that does not fit in int32_t, tripping
GenTreeIntCon::SetIconValue's FitsIn<int32_t>(value) assert during
'Morph - Global'.

Cast through int32_t first so the sign bit is preserved when widened to
ssize_t. Fixes #129099.

+++ src/coreclr/jit/importercalls.cpp
@@ -6408,7 +6408,11 @@ GenTree* Compiler::impPrimitiveNamedIntrinsic(NamedIntrinsic        intrinsic,
                 else
                 {
                     uint32_t cns1 = static_cast<uint32_t>(op1->AsIntConCommon()->IconValue());
-                    result        = gtNewIconNode(BitOperations::RotateLeft(cns1, cns2), baseType);
+                    // Sign-extend the unsigned fold result to int32_t so gtNewIconNode's
+                    // FitsIn<int32_t>(value) check for TYP_INT/TYP_UINT folds doesn't trip
+                    // on the high-bit-set case (e.g. RotateLeft(0xFFFFFFFFu, k)).
+                    result        = gtNewIconNode(
+                        static_cast<int32_t>(BitOperations::RotateLeft(cns1, cns2)), baseType);
                 }
                 break;
             }
@@ -6457,7 +6461,11 @@ GenTree* Compiler::impPrimitiveNamedIntrinsic(NamedIntrinsic        intrinsic,
                 else
                 {
                     uint32_t cns1 = static_cast<uint32_t>(op1->AsIntConCommon()->IconValue());
-                    result        = gtNewIconNode(BitOperations::RotateRight(cns1, cns2), baseType);
+                    // Sign-extend the unsigned fold result to int32_t so gtNewIconNode's
+                    // FitsIn<int32_t>(value) check for TYP_INT/TYP_UINT folds doesn't trip
+                    // on the high-bit-set case (e.g. RotateRight(0xFFFFFFFFu, k)).
+                    result        = gtNewIconNode(
+                        static_cast<int32_t>(BitOperations::RotateRight(cns1, cns2)), baseType);
                 }
                 break;
             }
